+++ Quanta/Source/Quanta/Graphics/Rendering/Texture/Texture2D.cpp
@@ -7,7 +7,7 @@
 
 namespace Quanta
 {
-    std::shared_ptr<Texture2D> Texture2D::Create(uint32_t width, uint32_t height)
+    std::shared_ptr<Texture2D> Texture2D::Create(size_t width, size_t height)
     {
         switch(GraphicsDevice::GetApi())
         {
